Updated ValidateCredentials to take an instance of AccountCredentials instead of the plain user name and password strings.
Added implementation of GetBookInfo, GetMovieInfo and GetAlbumInfo.

Added implementation of ValidateCredentials and CreateNewUser, but these are in the TestClass.cs to avoid breaking the code of these method that the Singaporians are currently testing.

Added LINQExtension class - Extensions for IEnumerable<T> for simple use in database query code.

Added the Util-class in InterfaceClasses.cs for convenient conversion to the enums MediaType and Rating.