Deleted the database proxy in BinaryCommunivcator
All methods of BinaryCommunicator has been implemented to reflect the removal of database clienting. Movie and Book has been tested, and works!

Added app.config to UserUploadControl (the same as the file in the BinaryCommunicator-project). Apparently this is needed, before the app.config-configurations of BinaryCommunicator is not accessible in it dll.