GetMedia web page: When a song is requested the method now checks that it has a rental of the album the song is a part of.
GetMedia web page: If a media file for a particular requested media does not exist in the database, a default file is now returned.

RentMedia web service: The method now checks if there already exists an active rental record of the specified media. If there does, the method throughs an exception.

Fixed the view details buttons in the three media playback windows.

When clicking the Rent-button on MediaSideBar, a confirmation box is now shown to the user.

Deleted code the initializes DetailedMediaListControl.

The rentals list can now be chosen to view either active or expired rentals.

Added additional message boxes for RentItMessageBox.